+++ src/app/models/maps/MyLine.java
@@ -88,7 +88,7 @@ public class MyLine implements Line {
         return lenght;
     }
 
-    public int getLenghtOfStreet(Street street){
+    public double getLenghtOfStreet(Street street){
         if (Math.abs(street.getCoordinates().get(0).getX()-street.getCoordinates().get(1).getX())>0){
             return (Math.abs(street.getCoordinates().get(0).getX()-street.getCoordinates().get(1).getX()));
         }
